Set to download installer first, then if there, remove files/folders and begin installing. Refactored

+++ extras/deb_package_files/home/kootnet_data/scripts/clean_upgrade_http.sh
@@ -7,29 +7,31 @@ if [[ $EUID != 0 ]]; then
 fi
 clear
 HTTP_SERVER="http://kootenay-networks.com"
-HTTP_FOLDER="/installers"
-DEB_INSTALLER="/KootnetSensors.deb"
-CONFIG_DIR="/etc/kootnet"
+HTTP_FOLDER="/installers/"
+DEB_INSTALLER="KootnetSensors.deb"
+CONFIG_DIR="/etc/kootnet/"
 INSTALL_TYPE="Standard"
 # Check for development switch
 if [[ "$1" == "dev" ]]; then
-  HTTP_FOLDER="/installers/dev"
+  HTTP_FOLDER="/installers/dev/"
   INSTALL_TYPE="Developmental"
 fi
 printf '\n-- %s Clean HTTP UPGRADE OR INSTALL --\n' "${INSTALL_TYPE}"
 printf '\nLeaving configurations and data, removing program folder and Python Virtual Environment\n'
-rm -f /tmp${DEB_INSTALLER} 2>/dev/null
-rm -R -f /home/kootnet_data/env 2>/dev/null
-rm -R -f /opt/kootnet-sensors 2>/dev/null
-rm -f /usr/share/applications/KootNet*.desktop 2>/dev/null
-rm -f /etc/systemd/system/Kootnet*.service 2>/dev/null
-rm -f /etc/systemd/system/SensorU*.service 2>/dev/null
-# SensorC* and SensorR* are for legacy service files (Much older versions)
-rm -f /etc/systemd/system/SensorC*.service 2>/dev/null
-rm -f /etc/systemd/system/SensorR*.service 2>/dev/null
-wget -O /tmp${DEB_INSTALLER} ${HTTP_SERVER}${HTTP_FOLDER}${DEB_INSTALLER}
-apt-get update
-apt-get -y --allow-downgrades --reinstall install /tmp${DEB_INSTALLER}
-# Save DateTime and Update type to file (Used in program to show last updated)
-date -u >${CONFIG_DIR}/last_updated.txt
-echo ' - Clean HTTP ' >>${CONFIG_DIR}/last_updated.txt
+mkdir /tmp 2>/dev/null
+rm -f /tmp/${DEB_INSTALLER} 2>/dev/null
+wget -O /tmp/${DEB_INSTALLER} ${HTTP_SERVER}${HTTP_FOLDER}${DEB_INSTALLER}
+# Make sure the installer file is there before deleting everything
+if [[ -f /tmp/${DEB_INSTALLER} ]]; then
+  rm -R -f /home/kootnet_data/env 2>/dev/null
+  rm -R -f /opt/kootnet-sensors 2>/dev/null
+  rm -f /usr/share/applications/KootNet*.desktop 2>/dev/null
+  rm -f /etc/systemd/system/Kootnet*.service 2>/dev/null
+  rm -f /etc/systemd/system/SensorU*.service 2>/dev/null
+  apt-get update
+  apt-get -y --reinstall --allow-downgrades install /tmp/${DEB_INSTALLER}
+  # Save DateTime and Update type to file (Used in program to show last updated)
+  date -u >${CONFIG_DIR}last_updated.txt
+  echo ' - Clean HTTP ' >>${CONFIG_DIR}last_updated.txt
+else
+  printf '\nHTTP Download Failed, Clean HTTP ' "${INSTALL_TYPE}" ' Upgrade Cancelled\n'
